Add documentation for viadot 2.0 with new repository structure (#929)
* 📝 Created new directory structure for references tab

* 📝 Added `Getting Started` section in docs

* 📝 Added `User Guide` section

* 📝 Refactored docs structure

* 📝 Added new user guide

* ✨ Added script to synchronize `.lock` files

* 📝 Added `Managing dependencies` section in docs

* 📝 Fixed typos in docs

* 📝 Improved tutorial about adding source and flows

* 📝 Removed `Manging dependecies section`

* 📝 Added flow and task referencies

* 📝 Updated link in documentation

* 📝 Updated docs in `user_guide/config_key.md`


* 📝 Updated docs in `user_guide/adding_source.md`


* 📝 Updated docs in `user_guide/adding_prefect_flow.md`


* 📝 Updated docs in `user_guide/adding_prefect_flow.md`


* 📝 Updated docs in `user_guide/adding_prefect_flow.md`


* 📝 Added description about `to_df()` in `adding_source.md`

* 📝 Improved docs in` adding_prefect_flow.md`

* 📝 Removed badges from `index.md`

* 📝 Added `Advanced Usage` section

* 📝 Moved docs about Rye into `CONTRIBUTING.md`

* 📝 Moved docker tutorial section form docs to `CONTRIBUTING.md`

* 📝 Updated `CONTRIBUTING.md`


* 📝 Removed Rye description from `CONTRIBUTING.md`

---------
