+++ docs/advanced_usage/docker_containers.md
@@ -1,0 +1,5 @@
+Currently there are tree available containers to build:
+
+- `viadot-lite` - It has installed default dependencies and supports only non-cloud-specific sources.
+- `viadot-azure` - It has installed default and viadot-azure dependencies. Supports Azure-based sources and non-cloud-specific ones.
+- `viadot-aws` - It has installed default and aws-azure dependencies. Supports AWS-based sources and non-cloud-specific ones.
