--- docs/references/api_sources.md
@@ -1,5 +1,0 @@
-# API Sources
-
-::: viadot.sources.uk_carbon_intensity.UKCarbonIntensity
-
-::: viadot.sources.cloud_for_customers.CloudForCustomers
