+++ docs/references/sources/api_sources.md
@@ -1,0 +1,17 @@
+# API Sources
+
+::: viadot.sources.uk_carbon_intensity.UKCarbonIntensity
+
+::: viadot.sources.cloud_for_customers.CloudForCustomers
+
+::: viadot.sources.exchange_rates.ExchangeRates
+
+::: viadot.sources.cloud_for_customers.CloudForCustomers
+
+::: viadot.sources.sap_rfc.SAPRFC
+
+::: viadot.sources.sharepoint.Sharepoint
+
+::: viadot.sources.genesys.Genesys
+
+::: viadot.sources.minio_source.MinIO
