+++ docs/references/sources/sql_sources.md
@@ -1,0 +1,17 @@
+# SQL Sources
+
+::: viadot.sources.base.Source
+
+::: viadot.sources.base.SQL
+
+::: viadot.sources.azure_data_lake.AzureDataLake
+
+::: viadot.sources.redshift_spectrum.RedshiftSpectrum
+
+::: viadot.sources.s3.S3
+
+::: viadot.sources.sqlite.SQLite
+
+::: viadot.sources.databricks.Databricks
+
+::: viadot.sources.trino_source.Trino
