--- docs/references/sql_sources.md
@@ -1,8 +1,0 @@
-# SQL Sources
-
-::: viadot.sources.base.Source
-::: viadot.sources.base.SQL
-
-::: viadot.sources.azure_data_lake.AzureDataLake
-
-::: viadot.sources.sqlite.SQLite
