--- docs/tutorials/databricks.md
@@ -1,5 +1,0 @@
-# Databricks
-
-## Credentials
-
-In order to retrieve values for the credentials, follow step 2 [here](https://docs.microsoft.com/en-us/azure/databricks/dev-tools/databricks-connect).
