--- docs/tutorials/supermetrics.md
@@ -1,13 +1,0 @@
-# How to pull Supermetrics data
-
-With `viadot` you have the opportunity to pull data from Supermetrics API and save it in parquet format on the Azure Data Lake. You can also load this data do Azure SQL Database.
-
-If you need more info about Supermetrics API, please visit https://supermetrics.com/docs/product-api-getting-started/
-
-## Pull data from Supermetrics
-
-Let's pull data from Supermetrics and save output as a parquet file on Azure Data Lake.
-
-To pull data from Supermetrics we will create flow basing on `SupermetricsToADLS`.
-
-Data types are automatically detected and mapped to meet Microsoft Azure SQL Database requirements. Schema json will be stored in the data lake (parquet_file_directory/schema)
